% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_back_missing_indvs.R
\name{add_back_missing_indvs}
\alias{add_back_missing_indvs}
\title{Adds individuals without scores, setting their CCI to 0}
\usage{
add_back_missing_indvs(icd_data, total_cci_scores)
}
\arguments{
\item{icd_data}{A data.frame with at least columns \code{ID}, the original
data.}

\item{total_cci_scores}{A data.frame with at least columsn \code{ID}, and
\code{CCI_score}.}
}
\value{
A data.frame with all individuals
}
\description{
If some individuals have not diagnosis in the exposure window
they will get exluded from the calculation. This function adds
them back setting their score to 0.
}
\author{
Kira E. Detrois
}
