% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_back_orig_ids.R
\name{add_back_orig_ids}
\alias{add_back_orig_ids}
\title{Add column with original IDs}
\usage{
add_back_orig_ids(cci_scores, icd_data)
}
\arguments{
\item{cci_scores}{A data.frame with at least column \code{ID_num},
and ideally column \code{score}.}

\item{icd_data}{A data.frame with at least columns \code{ID_num} and
\code{ID}.}
}
\value{
A data.frame with an added column with the original IDs
of the individuals.
}
\description{
Adding back the original IDs of the individuals mapped on to the
numeric IDs.
}
\details{
Since \code{\link[comorbidity]{comorbidity}} needs the IDs for each
individual to be numeric, the CCI scores need to be calculated, with
the original IDs mapped onto numeric ones. Technically, it could be
that the original IDs are already numeric but the mapping is done
regardless.
}
\author{
Kira E. Detrois
}
