% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_up_cci_scores.R
\name{add_up_cci_scores}
\alias{add_up_cci_scores}
\title{Adding up scores for different ICD-versions of the same
individual}
\usage{
add_up_cci_scores(cci_scores)
}
\arguments{
\item{cci_scores}{A data.frame with two columns: one called \code{score}
and one column defined by the variable \code{group_col}}
}
\value{
A data.frame with the scores added up for each individual.
This result is a tidy data.frame.
}
\description{
Adds up scores for different ICD-versions of the same
individual. Individuals are considered identical if they have the
same value in the column \code{ID_num}.
}
\examples{
cci_scores <- tibble::tibble(score = c(7, 0, 2, 3, 1),
                             ID_num = c(1, 1, 2, 3, 2))
add_up_cci_scores(cci_scores)

}
\author{
Kira E. Detrois
}
