% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cci.R
\name{calc_cci}
\alias{calc_cci}
\title{Calculates the Charlson Comorbidity Index based on ICD-codes}
\usage{
calc_cci(icd_data, exp_start = NA, exp_end = NA)
}
\arguments{
\item{icd_data}{A data.frame with at least columns \code{ID}, and
\code{PRIMARY_ICD}.}

\item{exp_start}{A numeric. Start of the exposure period. Can be
used to restrict the timeframe on which the index
should be calculated. In this case the data.frame
needs column \code{Event_age}.
In case it is a vector has to have the length of
the number of rows in \code{icd_data}.}

\item{exp_end}{A numeric. End of the exposure period.
Can be used to restrict the timeframe on which the index
should be calculated. In this case the data.frame
needs column \code{Event_age}.
In case it is a vector has to have the length of
the number of rows in \code{icd_data}.}
}
\value{
A tibble with columns \code{ID} and \code{score}.
Contains the charlson weighted comorbidity scores for each
individual.
}
\description{
Maps the ICD-codes onto the relevant conditions for the
charlson-deyo comorbidity index, using the function
\code{\link[comorbidity]{comorbidity}}.
Then calculates the scores, using the function
\code{\link[comorbidity]{score}}.
}
\details{
Can handle different ICD-versions also for a single patient by
adding up the results.
}
\author{
Kira E. Detrois
}
