% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cci_for_mult_exp_ages.R
\name{calc_cci_for_mult_exp_ages}
\alias{calc_cci_for_mult_exp_ages}
\title{Creates a named list of CCI score results for different exposure ages}
\usage{
calc_cci_for_mult_exp_ages(
  icd_data,
  exp_ages = c(20, 30, 40, 50),
  exp_len = 10
)
}
\arguments{
\item{icd_data}{A data.frame. For details see \link[ICCI]{calc_cci}}

\item{exp_ages}{An integer (vector). Ages at which exposure period
starts (in years).}

\item{exp_len}{An integer. Length of the exposure period
(in years).}
}
\value{
A list. The CCI score results data.frames for the different
exposure windows.
}
\description{
Creates a named list of CCI score results for different exposure ages
}
\author{
Kira E. Detrois
}
