% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_icd_ver_spec_ccis.R
\name{calc_icd_ver_spec_ccis}
\alias{calc_icd_ver_spec_ccis}
\title{Calculates the Charlson Comorbidity Index for a specific
ICD-version}
\usage{
calc_icd_ver_spec_ccis(long_data, icd_version)
}
\arguments{
\item{long_data}{A data.frame with at least columns \code{ID}, and
\code{PRIMARY_ICD}.}

\item{icd_version}{The ICD-version in INTERVENE format.
Can be either "10", "10CM", "9", or "9CM".}
}
\value{
A tibble with columns \code{ID} and \code{score}.
Contains the charlson weighted comorbidity scores for each
individual.
}
\description{
Maps the ICD-codes onto the relevant conditions for the
charlson-deyo comorbidity index, using the function
\code{\link[comorbidity]{comorbidity}}.
Then calculates the scores, using the function
\code{\link[comorbidity]{score}}.
}
\details{
Make sure that the ICD-codes in the data  all come from the same
ICD-version.
}
\author{
Kira E. Detrois
}
