% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mi.R
\name{calc_med}
\alias{calc_med}
\title{Calculates a medication score, based on ATC codes and associated weights}
\usage{
calc_med(atc_data, exp_start, exp_end)
}
\arguments{
\item{atc_data}{A data.frame with at least columns, \code{ID}, \code{ATC}, and \code{WEIGHT}}

\item{exp_start}{A numeric, or a data.frame with at least columns
\code{ID}, and \code{EXP_START}.
Start of the exposure period. Can be used to restrict
the timeframe on which the index should be calculated.
In this case the \code{icd_data} data.frame needs column
\code{Event_age}.
If the numeric is a vector has to have the length of
the number of rows in \code{icd_data}.}

\item{exp_end}{A numeric, or a data.frame with at least columns
\code{ID}, and \code{EXP_END}.
End of the exposure period. See \code{exp_start} docu.}
}
\description{
Calculates a medication score, based on ATC codes and associated weights
}
\author{
Kira E. Detrois
}
