% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_df.R
\name{create_test_atc_data}
\alias{create_test_atc_data}
\title{Creates test ATC data}
\usage{
create_test_atc_data(indv_ids, n_samples)
}
\arguments{
\item{indv_ids}{A character. IDs of the individuals.}

\item{n_samples}{A numeric. The number of samples to draw.}
}
\description{
Creates test ATC data
}
\author{
Kira E. Detrois
}
