% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_out_wrong_icd_vers.R
\name{filter_out_wrong_icd_vers}
\alias{filter_out_wrong_icd_vers}
\title{Filters out entries with incompatible ICD-version}
\usage{
filter_out_wrong_icd_vers(icd_data)
}
\arguments{
\item{icd_data}{A data.frame with at least column \code{ICD_version}.}
}
\value{
The data without rows with ICD-version entries other than "9", "9CM",
"10", or "10CM".
}
\description{
Currently, the CCI can only be automatically calculated on ICD-9 and 10.
}
\author{
Kira E. Detrois
}
