% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_long_data.R
\name{preprocess_icd_data}
\alias{preprocess_icd_data}
\title{Preprocesses the longitudinal data}
\usage{
preprocess_icd_data(icd_data, exp_start = NA, exp_end = NA)
}
\arguments{
\item{icd_data}{A data.frame with at least columns \code{ID}, and
\code{PRIMARY_ICD}.}

\item{exp_start}{A numeric. Start of the exposure period. Can be
used to restrict the timeframe on which the index
should be calculated. In this case the data.frame
needs column \code{Event_age}.
In case it is a vector has to have the length of
the number of rows in \code{icd_data}.}

\item{exp_end}{A numeric. End of the exposure period.
Can be used to restrict the timeframe on which the index
should be calculated. In this case the data.frame
needs column \code{Event_age}.
In case it is a vector has to have the length of
the number of rows in \code{icd_data}.}
}
\description{
Preprocesses the longidutdinal data by adding a numerical ID
to the data.frame and selecting the entries for the individuals
inside the selected exposure period.
}
\details{
Adds a numerical ID to the data.frame because the
\code{\link[comorbidity:comorbidity]{comordbidity}}
function expects the IDs to be numeric.
}
\author{
Kira E. Detrois
}
