% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_icd_ver_spec_ccis.R
\name{test_icd_ver_correct}
\alias{test_icd_ver_correct}
\title{Tests that there is only one ICD-version in the current data}
\usage{
test_icd_ver_correct(icd_data)
}
\arguments{
\item{icd_data}{A data.frame with at least columns \code{ID}, and
\code{PRIMARY_ICD}.}
}
\description{
Tests that there is only one ICD-version in the current data
}
\author{
Kira E. Detrois
}
